# docs/openapi_mobile_jeune.yaml
openapi: 3.0.3
info:
  title: Repartir - API Mobile Jeune
  version: 1.0.0
  description: |
    Spécification des endpoints nécessaires pour l'intégration Flutter côté Jeune.
    - Authentification via Bearer JWT
    - Horodatages au format ISO 8601 (UTC)
    - Pagination par défaut page=1, limit=20
    - Tri par défaut sur les listes par date desc (équivalent sort="-createdAt")
servers:
  - url: https://api.example.com
    description: Production
  - url: https://staging.api.example.com
    description: Staging
security:
  - bearerAuth: []
paths:
...
  /api/mentors:
    get:
      tags: [Mentors]
      summary: Liste publique des mentors
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: specialite
          schema: { type: string }
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Liste paginée de mentors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMentorDTO'
  /api/mentors/{mentorId}:
    get:
      tags: [Mentors]
      summary: Détail d'un mentor
      parameters:
        - in: path
          name: mentorId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Mentor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorDTO'
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
...
  responses:
...
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        totalItems: { type: integer }
        totalPages: { type: integer }
    PaginatedMentorDTO:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/MentorDTO' }
        pagination: { $ref: '#/components/schemas/Pagination' }
...
    MentorDTO:
      type: object
      properties:
        id: { type: string }
        nom: { type: string }
        prenom: { type: string }
        profession: { type: string }
        experienceAnnees: { type: integer }
        specialites:
          type: array
          items: { type: string }
        a_propos: { type: string }
        urlPhoto: { type: string, format: uri }
...